fix: Xray history will not be saved. Did refactor
